requirements.txt
praw
requests
Pillow
selenium
BeautifulSoup4
webdriver_manager
microsoft-bing-customimagesearch